+++ code/engine.vc2008/xrRender/xrRenderDX10/3DFluid/dx103DFluidRenderer.cpp
@@ -649,9 +649,9 @@ void dx103DFluidRenderer::CalculateLighting(const dx103DFluidData &FluidData, Fo
 		size
 		);
 
-	u32 iNumRenderables = m_lstRenderables.size();
+	size_t iNumRenderables = m_lstRenderables.size();
 	// Determine visibility for dynamic part of scene
-	for (u32 i=0; i<iNumRenderables; ++i)
+	for (size_t i=0; i<iNumRenderables; ++i)
 	{
 		ISpatial*	spatial		= m_lstRenderables[i];
